requirements.txt
pyaudio==0.2.14
openai-whisper==20231117
pynput==1.7.6
pvporcupine==3.0.2
numpy==1.24.3
torch==2.1.0
Pillow==10.1.0
webrtcvad==2.0.10
